Move feedback file watcher to it's own class
