+++ src/Razor/src/Microsoft.VisualStudio.LanguageServices.Razor/Logging/VisualStudioFeedbackLoggerProvider.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
-using System;
 using System.ComponentModel.Composition;
 using System.Threading;
 using Microsoft.CodeAnalysis.Razor.Logging;
